requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Redis
redis==5.2.0

# Validation
pydantic==2.10.2
pydantic-settings==2.6.1
email-validator==2.2.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Utils
httpx==0.27.2
python-json-logger==2.0.7
structlog==24.4.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
faker==30.8.2
factory-boy==3.3.1
